requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.8
Django==3.0.7
django-filter==2.2.0
djangorestframework==3.11.0
mysqlclient==1.4.6
paramiko==2.7.1
Pillow==7.0.0
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
unix==1.0
